requirements.txt
Django==3.2.8
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
pytz==2017.3
